+++ src/data/dummyData.js
@@ -5,27 +5,33 @@ export const data = {
   meta: {
     brainstrom: "브레인스톰",
   },
-  html: {
-    tree: "트리",
-    span: "<span>",
-    img: "<img>",
-    link: "<link>",
-    a: "<a>",
-    ul: "<ul>",
-    ol: "<ol>",
-    li: "<li>",
-    input: "<input>",
-    textarea: "<textarea>",
-    button: "<button>",
-    semantic: "시맨틱",
-    search: "검색",
-    footer: "<footer>",
-    header: "<header>",
-    nav: "<nav>",
-    main: "<main>",
-    id: "id",
-    class: "class",
-  },
+  html: [
+    {
+      tree: "트리",
+      id: "id",
+      class: "class",
+    },
+    {
+      span: "<span>",
+      img: "<img>",
+      link: "<link>",
+      a: "<a>",
+      ul: "<ul>",
+      ol: "<ol>",
+      li: "<li>",
+      input: "<input>",
+      textarea: "<textarea>",
+      button: "<button>",
+    },
+    {
+      semantic: "시맨틱",
+      search: "검색",
+      footer: "<footer>",
+      header: "<header>",
+      nav: "<nav>",
+      main: "<main>",
+    },
+  ],
   css: {
     flex: "flex",
     "flex-direction": "flex-direction",
@@ -58,10 +64,14 @@ export const data = {
     reset: "초기화",
   },
   js2: {},
-  test: {
-    flex: "flex",
-    "flex-direction": "flex-direction",
-    "justify-content": "justify-content",
-    "align-item": "align-item",
-  },
+  test: [
+    {
+      flex: "flex",
+    },
+    {
+      "flex-direction": "flex-direction",
+      "justify-content": "justify-content",
+      "align-item": "align-item",
+    },
+  ],
 };
